Enum relay_server::EnvelopeRepository

pub enum EnvelopeRepository {
    Memory(MemoryEnvelopeRepository),
    SQLite(SqliteEnvelopeRepository),
}

Represents different types of envelope repositories.

Variants

Memory(MemoryEnvelopeRepository)

In-memory envelope repository.

SQLite(SqliteEnvelopeRepository)

SQLite-based envelope repository.

Implementations

impl EnvelopeRepository

pub fn memory( memory_checker: MemoryChecker, ) -> Result<Self, EnvelopeBufferError>

Creates a new memory-based envelope repository.

pub async fn sqlite(config: &Config) -> Result<Self, EnvelopeBufferError>

Creates a new SQLite-based envelope repository.

pub async fn initialize(&mut self) -> InitializationState

Initializes the EnvelopeRepository and returns the initialization state.

pub async fn push( &mut self, project_key_pair: ProjectKeyPair, envelope: Box<Envelope>, ) -> Result<(), EnvelopeBufferError>

Pushes an envelope to the repository for the given project key pair.

pub async fn peek( &mut self, project_key_pair: ProjectKeyPair, ) -> Result<Option<&Envelope>, EnvelopeBufferError>

Peeks at the next envelope for the given project key pair without removing it.

pub async fn pop( &mut self, project_key_pair: ProjectKeyPair, ) -> Result<Option<Box<Envelope>>, EnvelopeBufferError>

Pops and returns the next envelope for the given project key pair.

pub async fn flush(&mut self) -> bool

Flushes the [Envelope]s in the EnvelopeRepository.

pub fn has_store_capacity(&self) -> bool

Returns true when there is space to store more [Envelope]s, false otherwise.

pub async fn store_total_count(&self) -> u64

Returns the total count of [Envelope]s in the store.

pub fn name(&self) -> &'static str

Returns the string representation of the EnvelopeRepository’s strategy.