relay_server
Function run

pub fn run(config: Config) -> Result<()>

Runs a relay web server and spawns all internal worker threads.

This effectively boots the entire server application. It blocks the current thread until a shutdown signal is received or a fatal error happens. Behavior of the server is determined by the config passed into this funciton.