relay_server

Module managed

Managed containers and utilities to ensure outcomes are correctly emitted.

Structs

InvalidProcessingGroupType

Error emitted when converting a ManagedEnvelope and a processing group into a TypedEnvelope.

Managed

The Managed wrapper ensures outcomes are correctly emitted for the contained item.

ManagedEnvelope

Tracks the lifetime of an [Envelope] in Relay.

RecordKeeper

A record keeper makes sure modifications done on a Managed item are all accounted for correctly.

Rejected

A wrapper type which ensures outcomes have been emitted for an error.

Split

Iterator returned by Managed::split.

TypedEnvelope

A wrapper for ManagedEnvelope with assigned processing group type.

Enums

ItemAction

Represents the decision on whether or not to keep an envelope item.

Traits

Counted

A counted item.

ManagedResult

Extension trait for Results, which adds convenience functions to work with errors and Managed.

OutcomeError

An error which can be extracted into an outcome.

Type Aliases

Quantities

A list of data categories and amounts.