relay_server::managed

Struct Managed

pub struct Managed<T: Counted> { /* private fields */ }

The Managed wrapper ensures outcomes are correctly emitted for the contained item.

Implementations

impl Managed<Box<Envelope>>

pub fn from_envelope( envelope: Box<Envelope>, outcome_aggregator: Addr<TrackOutcome>, ) -> Self

Creates a managed instance from an unmanaged envelope.

impl<T: Counted> Managed<T>

pub fn with_meta_from(envelope: &ManagedEnvelope, value: T) -> Self

Creates new Managed instance with the provided value and metadata from a ManagedEnvelope.

The Managed instance, inherits all metadata from the passed ManagedEnvelope, like received time or scoping.

pub fn wrap<S>(&self, other: S) -> Managed<S>
where S: Counted,

Creates another Managed instance, with a new value but shared metadata.

pub fn received_at(&self) -> DateTime<Utc>

Original received timestamp.

pub fn scoping(&self) -> Scoping

Scoping information stored in this context.

pub fn scope(&mut self, scoping: Scoping)

Updates the scoping stored in this context.

Special care has to be taken when items contained in the managed instance also store a scoping. Such a scoping will not be updated.

Conversions between Managed<Box<Envelope>> and ManagedEnvelope transfer the scoping correctly.

See also: ManagedEnvelope::scope.

pub fn split_once<F, S, U>(self, f: F) -> (Managed<S>, Managed<U>)
where F: FnOnce(T, &mut RecordKeeper<'_>) -> (S, U), S: Counted, U: Counted,

Splits Self into two other Managed items.

The two resulting managed instances together are expected to have the same outcomes as the original instance.. Since splitting may introduce a new type of item, which some of the original quantities are transferred to, there may be new additional data categories created.

pub fn split<F, I, S>(self, f: F) -> Split<I::IntoIter, I::Item>
where F: FnOnce(T) -> I, I: IntoIterator<Item = S>, S: Counted,

Splits Self into a variable amount if individual items.

Useful when the current instance contains multiple items of the same type and must be split into individually managed items.

pub fn split_with_context<F, I, S, C>( self, f: F, ) -> (Split<I::IntoIter, I::Item>, C)
where F: FnOnce(T) -> (I, C), I: IntoIterator<Item = S>, S: Counted,

Splits Self into a variable amount if individual items.

Like Self::split but also allows returning an untracked context, a way of returning additional data when deconstructing the original item.

pub fn retain<S, I, U, E>(&mut self, select: S, retain: U)
where S: FnOnce(&mut T) -> &mut Vec<I>, I: Counted, U: FnMut(&mut I, &mut RecordKeeper<'_>) -> Result<(), E>, E: OutcomeError,

Filters individual items and emits outcomes for them if they are removed.

This is particularly useful when the managed instance is a container of individual items, which need to be processed or filtered on a case by case basis.

Examples:

struct Items {
    items: Vec<Item>,
}

fn process_items(items: &mut Managed<Items>, ctx: Context<'_>) {
    items.retain(|items| &mut items.items, |item, _| process(item, ctx));
}

fn process(item: &mut Item, ctx: Context<'_>) -> Result<(), Error> {
    todo!()
}

pub fn retain_with_context<S, C, I, U, E>(&mut self, select: S, retain: U)
where for<'a> S: FnOnce(&'a mut T) -> (&'a mut Vec<I>, &'a C), I: Counted, U: FnMut(&mut I, &C, &mut RecordKeeper<'_>) -> Result<(), E>, E: OutcomeError,

Filters individual items and emits outcomes for them if they are removed.

Like Self::retain, but it allows for an additional context extracted from the managed object passed to the retain function.

Examples:

struct Items {
    ty: String,
    items: Vec<Item>,
}

fn process_items(items: &mut Managed<Items>, ctx: Context<'_>) {
    items.retain_with_context(|items| (&mut items.items, &items.ty), |item, ty, _| process(item, ty, ctx));
}

fn process(item: &mut Item, ty: &str, ctx: Context<'_>) -> Result<(), Error> {
    todo!()
}

pub fn map<S, F>(self, f: F) -> Managed<S>
where F: FnOnce(T, &mut RecordKeeper<'_>) -> S, S: Counted,

Maps a Managed<T> to Managed<S> by applying the mapping function f.

Like Self::try_map but not fallible.

pub fn try_map<S, F, E>(self, f: F) -> Result<Managed<S>, Rejected<E::Error>>
where F: FnOnce(T, &mut RecordKeeper<'_>) -> Result<S, E>, S: Counted, E: OutcomeError,

Maps a Managed<T> to Managed<S> by applying the mapping function f.

The mapping function gets access to a RecordKeeper, to emit outcomes for partial discards.

If the mapping function returns an error, the entire (original) Self is rejected, no partial outcomes are emitted.

pub fn modify<F>(&mut self, f: F)
where F: FnOnce(&mut T, &mut RecordKeeper<'_>),

Gives mutable access to the contained value to modify it.

Like Self::try_modify but not fallible.

pub fn try_modify<F, E>(&mut self, f: F) -> Result<(), Rejected<E::Error>>
where F: FnOnce(&mut T, &mut RecordKeeper<'_>) -> Result<(), E>, E: OutcomeError,

Gives mutable access to the contained value to modify it.

The modifying function gets access to a RecordKeeper, to emit outcomes for partial discards.

If the modifying function returns an error, the entire (original) Self is rejected, no partial outcomes are emitted.

pub fn accept<F, S>(self, f: F) -> S
where F: FnOnce(T) -> S,

Accepts the item of this managed instance.

This should be called if the item has been or is about to be accepted by the upstream, which means that the responsibility for logging outcomes has been moved. This function will not log any outcomes.

Like Self::try_accept, but infallible.

pub fn try_accept<F, S, E>(self, f: F) -> Result<S, Rejected<E::Error>>
where F: FnOnce(T) -> Result<S, E>, E: OutcomeError,

Accepts the item of this managed instance.

This should be called if the item has been or is about to be accepted by the upstream.

Outcomes are only emitted when the accepting closure returns an error, which means that in the success case the responsibility for logging outcomes has been moved to the caller/upstream.

pub fn internal_error(&self, reason: &'static str) -> Rejected<()>

Rejects the entire Managed instance with an internal error.

Internal errors should be reserved for uses where logical invariants are violated. Cases which should never happen and always indicate a logical bug.

This function will panic in debug builds, but discard the item with an internal discard reason in release builds.

pub fn reject_err<E>(&self, error: E) -> Rejected<E::Error>
where E: OutcomeError,

Rejects the entire Managed instance.

impl<T: Counted> Managed<Option<T>>

pub fn transpose(self) -> Option<Managed<T>>

Turns a managed option into an optional Managed.

impl<L: Counted, R: Counted> Managed<Either<L, R>>

pub fn transpose(self) -> Either<Managed<L>, Managed<R>>

Turns a managed Either into an Either of Managed.