relay_server::managed

Trait OutcomeError

pub trait OutcomeError {
    type Error;

    // Required method
    fn consume(self) -> (Option<Outcome>, Self::Error);
}

An error which can be extracted into an outcome.

Required Associated Types

type Error

Produced error, without attached outcome.

Required Methods

fn consume(self) -> (Option<Outcome>, Self::Error)

Consumes the error and returns an outcome and Self::Error.

Returning a None outcome should discard the item(s) silently.