Trait relay_server::EnvelopeStack

pub trait EnvelopeStack: Send + Debug {
    type Error: Debug;

    // Required methods
    fn push(
        &mut self,
        envelope: Box<Envelope>,
    ) -> impl Future<Output = Result<(), Self::Error>>;
    fn peek(
        &mut self,
    ) -> impl Future<Output = Result<Option<&Envelope>, Self::Error>>;
    fn pop(
        &mut self,
    ) -> impl Future<Output = Result<Option<Box<Envelope>>, Self::Error>>;
    fn flush(self) -> Vec<Box<Envelope>>;
}

A stack-like data structure that holds [Envelope]s.

Required Associated Types

type Error: Debug

The error type that is returned when an error is encountered during reading or writing the EnvelopeStack.

Required Methods

fn push( &mut self, envelope: Box<Envelope>, ) -> impl Future<Output = Result<(), Self::Error>>

Pushes an [Envelope] on top of the stack.

fn peek( &mut self, ) -> impl Future<Output = Result<Option<&Envelope>, Self::Error>>

Peeks the [Envelope] on top of the stack.

fn pop( &mut self, ) -> impl Future<Output = Result<Option<Box<Envelope>>, Self::Error>>

Pops the [Envelope] on top of the stack.

fn flush(self) -> Vec<Box<Envelope>>

Persists all envelopes in the EnvelopeStacks to external storage, if possible, and consumes the stack provider.

Object Safety

This trait is not object safe.